requirements.txt
pillow
tqdm
notebook
ipywidgets
streamlit
cohere
openai
python-dotenv
faiss-cpu
pymupdf
